perf-test/requirements.txt
# Performance Testing Framework Dependencies

# Core dependencies (should already be available from main project)
rich>=13.0.0
loguru>=0.7.0
pyyaml>=6.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Data analysis and visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# GPU monitoring (optional)
pynvml>=11.5.0

# Additional utilities
scipy>=1.10.0
scikit-learn>=1.3.0